requirements.txt
nltk==3.2.5
tensorboard_logger==0.0.4
records==0.5.2
more_itertools==3.2.0
torchtext==0.2.0
torch==0.2.0.post3
six==1.11.0
Babel==2.5.1
tabulate==0.8.1
tqdm==4.19.8
numpy==1.14.2
forked_path==0.2.3
scikit_learn==0.19.1
stanza==0.3
